+++ packages/paste-core/components/label/src/index.tsx
@@ -1,1 +1,2 @@
 export * from './Label';
+export * from './RequiredDot';
